Move code to widget_pages
