+++ .gitignore
@@ -1,1 +1,2 @@
 widget_pages/__pycache__
+users.txt
